rename enum to utils, time_since_previous_action to time_delta, and keys_pressed to keys
